Exercise-3 and new files

+++ exercise-2/heapSort.c
@@ -1,5 +1,6 @@
 #include <stdio.h>
 #include <stdlib.h>
+#include <stdint.h>
 
 #include "heapSort.h"
 
@@ -17,9 +18,9 @@
  * Entrada: tamanho máximo da heap.
  * Saída..: Heap.
  */
-Heap* create_heap (unsigned int size) {
+Heap* create_heap (uint64_t size) {
   Heap* heap = (Heap*) malloc(sizeof(Heap));
-  heap -> table = (int*) malloc(sizeof(int) * (size + 1));
+  heap -> table = (uint64_t*) malloc(sizeof(uint64_t) * (size + 1));
   heap -> size = size;
   heap -> length = 0;
 
@@ -31,7 +32,7 @@ Heap* create_heap (unsigned int size) {
  * Entrada: indice do pai.
  * Saída..: indice do filho esquerdo.
  */
-int left_son (int dad) {
+uint64_t left_son (uint64_t dad) {
   return 2 * dad;
 }
 
@@ -41,7 +42,7 @@ int left_son (int dad) {
  * Entrada: indice do pai.
  * Saída..: indice do filho direito.
  */
-int right_son (int dad) {
+uint64_t right_son (uint64_t dad) {
   return 2 * dad + 1;
 }
 
@@ -49,10 +50,10 @@ int right_son (int dad) {
  * Entrada: array, tamanho do array
  * Saída..: exibe array
  */
-void show_sorted_array (int* array, int size) {
+void show_sorted_array (uint64_t* array, uint64_t size) {
   printf("Sorted Array: [");
-  for (int i = 1; i <= size; i++) {
-    printf("%d",array[i]);
+  for (uint64_t i = 1; i <= size; i++) {
+    printf("%lu",array[i]);
     if (i != size) printf(", ");
   }
   printf("]\n");
@@ -68,8 +69,8 @@ void free_memory (Heap* heap) {
  * Entrada: indice do elemento 1, indice do elemento 2, heap
  * Saída..: ---------------
  */
-void swap (int i, int j, Heap* heap) {
-  int aux = heap -> table[i];
+void swap (uint64_t i, uint64_t j, Heap* heap) {
+  uint64_t aux = heap -> table[i];
   heap -> table[i] = heap -> table[j];
   heap -> table[j] = aux;
 }
@@ -80,8 +81,8 @@ void swap (int i, int j, Heap* heap) {
  * Saída ......: ---------
  * Complexidade: O(log n)
  */
-void down (int dad, Heap* heap) {
-  int j = left_son(dad);
+void down (uint64_t dad, Heap* heap) {
+  uint64_t j = left_son(dad);
 
   if (j <= heap -> length) {
     if (j < heap -> length) {
@@ -103,7 +104,7 @@ void down (int dad, Heap* heap) {
  * Saída..: heap
  * Complexidade: O(n)
  */
-Heap* build_heap (int* array, int size) {
+Heap* build_heap (uint64_t* array, uint64_t size) {
   Heap* heap = create_heap(size);
   heap -> length = size;
   
@@ -111,7 +112,7 @@ Heap* build_heap (int* array, int size) {
   heap -> table = array;
   (heap -> table)--; 
   
-  for (int i = (heap->length) / 2; i >= 1; i--)
+  for (uint64_t i = (heap->length) / 2; i >= 1; i--)
     down(i, heap);
 
   return heap;
@@ -123,10 +124,10 @@ Heap* build_heap (int* array, int size) {
  * Saída..: vetor ordenado
  * Complexidade: O(n-1) * O(log n) = O(n * log n)
  */
-int* heapSort (int* array, int size) {
+uint64_t* heapSort (uint64_t* array, uint64_t size) {
   Heap* heap = build_heap(array, size); // O(n)
 
-  for (int i = heap -> length; i >= 2; i--) { // O(n-1)
+  for (uint64_t i = heap -> length; i >= 2; i--) { // O(n-1)
     swap (1, i, heap);
     heap -> length = heap -> length - 1;
     down (1, heap); // O(log n)
